[JENKINS-39109] Add a configuration option for what label to use for docker agents
This adds a folder property and a global configuration where the administrator,
or someone configuring a folder, can specify what label expression to use with
a pipeline with agent docker: "" and omitting the label.
The folder property can be specified on any folder in the hierarchy and the closest
parent to the run is consulted first and then up the tree, lastly the global configuration
is consulted for a value.
An extension point is also introduced with specified ordinals for the existing extensions
with ample room to spare in between in case something wants to jump in between.
Lastly a step that returns the value of the extension point that first returns a non null value.