Get rid of declarativeProps step and fix context for docker
The context fix is the easy one. The declarativeProps...well, why is
it even a step? We're basically just using it internally, we can
whitelist it, we can avoid the log spam of "[Pipeline]
declarativeProps" showing up a bunch, we can avoid cluttering the flow
graph... Yeah.

Also cleaned up the fact that I somehow bungled all the things and
lost the inheritance of AbstractDockerAgent in DockerPipeline.